requirements.txt
docx-mailmerge==0.5.0
lxml==5.1.0
numpy==1.26.4
pandas==2.20
python-dateutil==2.8.2
openpyxl==3.1.2
